Move core headers into dust tree
