+++ inst/include/dust/utils.hpp
@@ -95,4 +95,18 @@ private:
 }
 }
 
+namespace mode {
+
+template<typename T>
+T square(T x) {
+  return x * x;
+}
+
+template<typename T>
+T clamp(T x, T min, T max) {
+  return std::max(std::min(x, max), min);
+}
+
+}
+
 #endif
